restructured some directories
