+++ src/main/java/space/parzival/shardbot/commands/music/Bassboost.java
@@ -15,7 +15,7 @@ import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import space.parzival.shardbot.exceptions.CommandExecutionException;
-import space.parzival.shardbot.music.AudioControl;
+import space.parzival.shardbot.modules.music.AudioControl;
 import space.parzival.shardbot.types.Command;
 import space.parzival.shardbot.types.RichEmbedBuilder;
