+++ src/main/java/space/parzival/shardbot/commands/music/Play.java
@@ -19,8 +19,8 @@ import net.dv8tion.jda.api.entities.VoiceChannel;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.InteractionHook;
 import space.parzival.shardbot.exceptions.CommandExecutionException;
-import space.parzival.shardbot.music.AudioControl;
-import space.parzival.shardbot.music.TrackScheduler;
+import space.parzival.shardbot.modules.music.AudioControl;
+import space.parzival.shardbot.modules.music.TrackScheduler;
 import space.parzival.shardbot.types.Command;
 import space.parzival.shardbot.types.RichEmbedBuilder;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
