+++ src/main/java/space/parzival/shardbot/modules/music/AudioPlayerSendHandler.java
@@ -1,4 +1,4 @@
-package space.parzival.shardbot.music;
+package space.parzival.shardbot.modules.music;
 
 import java.nio.ByteBuffer;
